[AO] Experimental implementation of GTAO
This is an experimental implementation of GTAO (Ground Truth Ambient
Occlusion) for the Post Processing Stack. This only implements the GTAO
integrator without the spatio-temporal denoising technique in the
original paper. It means that the implementation is quite slow at the
moment -- it's several times slower than the current AO implementation
in the stack.